requirements.txt
flask
gunicorn
pandas
numpy
matplotlib
scikit-learn
mysql-connector-python
python-dotenv